Pin numpy<2.0.0 (#178)
* Add new Python versions to testing matrix

* Pin numpy in environment-ci.yml

There are still a few incompatibilities with numpy 2.0.0, so we'll pin
the previous version for the moment. In particular, xESMF regridders
can't handle large grids with a number of points that overflows a
32-bit integer.

* Pin numpy<2.0.0 in pyproject.toml

* Remove Python 3.11 from testing

For some reason it seems to hang indefinitely. We'll just disable it
for the moment (3.12 works, anyway).